+++ packages/stacked/lib/src/code_generation/stacked_locator_annotations.dart
@@ -6,22 +6,28 @@ class DependencyRegistration {
   /// An abstracted class type of service to register
   final Type asType;
 
-  const DependencyRegistration({this.classType, this.asType});
+  final Function resolveUsing;
+
+  const DependencyRegistration(
+      {this.classType, this.asType, this.resolveUsing});
 }
 
 /// Registers the type passed in as a singleton instance in the get_it locator
 class Singleton extends DependencyRegistration {
-  const Singleton({Type classType, Type asType}) : super(classType: classType, asType: asType);
+  const Singleton({Type classType, Type asType, Function resolveUsing})
+      : super(classType: classType, asType: asType, resolveUsing: resolveUsing);
 }
 
 /// Registers the type passed in as a LazySingleton instance in the get_it locator
 class LazySingleton extends DependencyRegistration {
-  const LazySingleton({Type classType, Type asType}) : super(classType: classType, asType: asType);
+  const LazySingleton({Type classType, Type asType, Function resolveUsing})
+      : super(classType: classType, asType: asType, resolveUsing: resolveUsing);
 }
 
 /// Registers the type passed in as a Factory in the get_it locator
 class Factory extends DependencyRegistration {
-  const Factory({Type classType, Type asType}) : super(classType: classType, asType: asType);
+  const Factory({Type classType, Type asType})
+      : super(classType: classType, asType: asType);
 }
 
 /// Registers the type passed in to be presolved using the function passed in