+++ packages/stacked/lib/stacked_annotations.dart
@@ -1,3 +1,3 @@
 export 'src/code_generation/stacked_router_annotations.dart';
-export 'src/code_generation/stacked_getit_annotations.dart';
+export 'src/code_generation/stacked_locator_annotations.dart';
 export 'src/code_generation/forms/stacked_form_annotations.dart';